[MOV-17] refactor enums
- Added a trait for common methods like getAllValues().
- Removed custom Enum exceptions; replaced them with the standard ValueError.
- Refactoring:
  - Moved common methods from ArticleStatusEnum and SectionEnum to the new trait.
  - Updated tests to align with these changes.
