--- src/app/Exceptions/Enum/ArticleStatusEnumInvalidException.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\Exceptions\Enum;
-
-class ArticleStatusEnumInvalidException extends EnumException
-{
-    protected $message = 'Invalid value for ArticleStatusEnum';
-}
