--- src/app/Exceptions/Enum/SectionEnumInvalidException.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\Exceptions\Enum;
-
-class SectionEnumInvalidException extends EnumException
-{
-    protected $message = 'Invalid value for SectionEnum';
-}
