+++ src/app/Traits/EnumHelper.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Traits;
+
+trait EnumHelper
+{
+    /**
+     * Return all values
+     *
+     * @return array
+     */
+    public static function getAllValues(): array
+    {
+        return array_map(static fn ($case) => $case->value, self::cases());
+    }
+}
